fix(payday): resolve LSP type diagnostics across source and tests
- Initialize er_ni_saving to fix possibly unbound variable
- Wrap float args with round() for int parameters
- Widen parameter types to float where callers pass float
- Add type annotation to template dict in generate_template
- Add assert guards for Optional fields in test files
- Fix callable -> Callable type hint in cli.py

+++ payday/dividend_tax.py
@@ -19,7 +19,7 @@ _HIGHER_BAND_WIDTH = HIGHER_RATE_BAND_LIMIT - BASIC_RATE_BAND_LIMIT  # 74,870
 
 
 def _compute_dividend_tax_bands(
-    dividends: int, total_employment: int, ani: int
+    dividends: int, total_employment: float, ani: int
 ) -> dict:
     """Core dividend tax computation for a given dividend amount.
     Returns a dict with all band and tax values plus allowance breakdown.
